+++ iterableapi/src/main/java/com/iterable/iterableapi/IterableEmbeddedManager.kt
@@ -13,7 +13,7 @@ public class IterableEmbeddedManager : IterableActivityMonitor.AppStateCallback 
 
     // region variables
     private var localPlacementMessagesMap = mutableMapOf<Long, List<IterableEmbeddedMessage>>()
-    private var placementIds = mutableListOf<Long>()
+    private var localPlacementIds = mutableListOf<Long>()
 
     private var updateHandleListeners = mutableListOf<IterableEmbeddedUpdateHandler>()
     private var iterableApi: IterableApi
@@ -77,15 +77,19 @@ public class IterableEmbeddedManager : IterableActivityMonitor.AppStateCallback 
     }
 
     fun getPlacementIds(): List<Long> {
-        return placementIds
+        return localPlacementIds
+    }
+
+    fun syncMessages() {
+        syncMessages(emptyArray<Long>())
     }
 
     //Network call to get the embedded messages
-    fun syncMessages() {
+    fun syncMessages(placementIds: Array<Long>) {
         if (iterableApi.config.enableEmbeddedMessaging) {
             IterableLogger.v(TAG, "Syncing messages...")
 
-            IterableApi.sharedInstance.getEmbeddedMessages(SuccessHandler { data ->
+            IterableApi.sharedInstance.getEmbeddedMessages(placementIds, { data ->
                 IterableLogger.v(TAG, "Got response from network call to get embedded messages")
                 try {
                     val previousPlacementIds = getPlacementIds()
@@ -135,7 +139,7 @@ public class IterableEmbeddedManager : IterableActivityMonitor.AppStateCallback 
                     }
 
                     //store placements from payload for next comparison
-                    placementIds = currentPlacementIds
+                    localPlacementIds = currentPlacementIds
 
                 } catch (e: JSONException) {
                     IterableLogger.e(TAG, e.toString())